LPS-56703 Move upgrades to its own module
