+++ modules/apps/blogs/blogs-web/src/com/liferay/blogs/web/upgrade/BlogsWebUpgrade.java
@@ -14,13 +14,15 @@
 
 package com.liferay.blogs.web.upgrade;
 
-import com.liferay.blogs.web.constants.BlogsPortletKeys;
+import com.liferay.blogs.web.upgrade.v1_0_0.UpgradePortletId;
+import com.liferay.blogs.web.upgrade.v1_0_0.UpgradePortletSettings;
 import com.liferay.portal.kernel.exception.PortalException;
+import com.liferay.portal.kernel.settings.SettingsFactory;
 import com.liferay.portal.kernel.upgrade.UpgradeProcess;
 import com.liferay.portal.service.ReleaseLocalService;
-import com.liferay.portal.upgrade.util.UpgradePortletId;
 
-import java.util.Collections;
+import java.util.ArrayList;
+import java.util.List;
 
 import javax.servlet.ServletContext;
 
@@ -45,27 +47,24 @@ public class BlogsWebUpgrade {
 	protected void setServletContext(ServletContext servletContext) {
 	}
 
+	@Reference(unbind = "-")
+	protected void setSettingsFactory(SettingsFactory settingsFactory) {
+		_settingsFactory = settingsFactory;
+	}
+
 	@Activate
 	protected void upgrade() throws PortalException {
-		UpgradePortletId upgradePortletId = new UpgradePortletId() {
+		List<UpgradeProcess> upgradeProcesses = new ArrayList<>();
 
-			@Override
-			protected String[][] getRenamePortletIdsArray() {
-				return new String[][] {
-					new String[] {"33", BlogsPortletKeys.BLOGS},
-					new String[] {"115", BlogsPortletKeys.BLOGS_AGGREGATOR},
-					new String[] {"161", BlogsPortletKeys.BLOGS_ADMIN}
-				};
-			}
+		upgradeProcesses.add(new UpgradePortletId());
 
-		};
+		upgradeProcesses.add(new UpgradePortletSettings(_settingsFactory));
 
 		_releaseLocalService.updateRelease(
-			"com.liferay.blogs.web",
-			Collections.<UpgradeProcess>singletonList(upgradePortletId), 1, 1,
-			false);
+			"com.liferay.blogs.web", upgradeProcesses, 1, 1, false);
 	}
 
 	private ReleaseLocalService _releaseLocalService;
+	private SettingsFactory _settingsFactory;
 
 }
