+++ modules/apps/blogs/blogs-web/src/com/liferay/blogs/web/upgrade/v1_0_0/UpgradePortletSettings.java
@@ -1,0 +1,39 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.blogs.web.upgrade.v1_0_0;
+
+import com.liferay.blogs.web.constants.BlogsPortletKeys;
+import com.liferay.portal.kernel.settings.SettingsFactory;
+import com.liferay.portal.util.PortletKeys;
+import com.liferay.portlet.blogs.util.BlogsConstants;
+
+/**
+ * @author Sergio González
+ */
+public class UpgradePortletSettings
+	extends com.liferay.portal.upgrade.v7_0_0.UpgradePortletSettings {
+
+	public UpgradePortletSettings(SettingsFactory settingsFactory) {
+		super(settingsFactory);
+	}
+
+	@Override
+	protected void doUpgrade() throws Exception {
+		upgradeMainPortlet(
+			BlogsPortletKeys.BLOGS, BlogsConstants.SERVICE_NAME,
+			PortletKeys.PREFS_OWNER_TYPE_GROUP, true);
+	}
+
+}
